Add SettingsScreen

+++ core/src/main/java/com/hyperion/template/MyGdxGame.java
@@ -16,8 +16,8 @@ public class MyGdxGame extends Game {
 
     public static final String TITLE = "MyGdxGame";
 
-    public static final short WIDTH = 1280;
-    public static final short HEIGHT = 720;
+    public static final int WIDTH = 1280;
+    public static final int HEIGHT = 720;
 
     private static String version;
